+++ packages/serve-sim/src/client/utils/panel-widths.ts
@@ -1,0 +1,3 @@
+export const PANEL_WIDTH = 320;
+export const DEVTOOLS_PANEL_WIDTH = 760;
+export const GRID_PANEL_WIDTH = 720;
